Better tf parsing to GPU
Using the position of the nodes relative to the widget, making the x axis actually be the horizontal position of the nodes instead of just a visualization. There is a neat way to calculate the t from the resulting point, but I couldn't figure that out and instead brute-force this by approximation using average of tens.